Update deploy scripts to actually work
As tuned in #123, this appears to actually work. The catch is that all of my testing was done with a single script that also retrieves the source, whereas for CodeDeploy, it's necessary to break it up into pre-deploy and post-deploy steps. CodeDeploy is also weirdly clueless as to whether this is a deploy that's running an update to an existing server or performing an installation to a fresh server. So I had to add conditionals, to avoid reconfiguring existing servers, and the best condition that I could think of was to see if the site already existed on the server.