Load Sphinx library via Composer
I did this in `master`, but apparently I haven't brought it over to `deploy`, I think because it's part of a bunch of PHP 7 upgrade work. So, here I set up inclusion via Composer, but don't actually at on it yet by modifying `search.php`, which still uses the manually-installed library.