Fix broken tooltips
Moving from JSON-P to CORS broke the tooltips. Turned out that the deprecated Ajax method on qTip was the culprit. I moved to a modern method (wrapping Ajax in an anonymous function), and rewrote all tooltip queries to use that, and now everything works great.

Also, I fixed the legislator elected date on their tooltips.