Change how the database export works
The long list of tables is hitting mysqldump's hard limit on how long the list can be, which, based on a review of the source code for mysqldump, seems to assume will only list a single table, despite allowing multiple tables. So now, instead, mysqldump is run once per table.