Select the most recent sub-year session
Generally, there's no duplication of bill numbers between sessions. But 2021 blew that up, with a special session that's really a continuation of the main session. So we have two copies of each bill, one of which is dead, and the other which is very much alive. Deal with this by ordering by session ID.