Remove race selector
In the years since I wrote this, three things have changed. The first is that the racial identities of many legislators have become a lot more representative of Virginia. The second is that there are increasing numbers of legislators whose racial identities cannot be reduced to a label. The third is that the racial descriptors used by the house (the senate doesn't provide any) have become free-form—which is good. Anyway, this old approach doesn't work, and I need to rethink how I gather and present that data. At the moment, the data isn't accurate, in that it's too reductive and doesn't reflect people's identities. So best to remove it from here, the only place on the site where it's shown.